requirements.txt
# Image Converter — Dependencies
pillow
pillow-heif
PySide6
